+++ CHANGELOG.md
@@ -1,4 +1,42 @@
 
+Version 2.0.0-RC6 (August 18th, 2016)
+=====================================
+
+  * New: Add ability to set multiple Callback instances.
+  * New: Add plugin API.
+
+```java
+class LoggingPlugin implements Plugin {
+  @Override public void configure(Analytics.Builder builder) {
+    builder.log(new Log() {
+      @Override public void print(Level level, String format, Object... args) {
+        System.out.println(level + ":\t" + String.format(format, args));
+      }
+
+      @Override public void print(Level level, Throwable error, String format, Object... args) {
+        System.out.println(level + ":\t" + String.format(format, args));
+        System.out.println(error);
+      }
+    });
+
+    builder.callback(new Callback() {
+      @Override public void success(Message message) {
+        System.out.println("Uploaded " + message);
+      }
+
+      @Override public void failure(Message message, Throwable throwable) {
+        System.out.println("Could not upload " + message);
+        System.out.println(throwable);
+      }
+    });
+  }
+}
+
+final Analytics analytics = Analytics.builder(writeKey) //
+        .plugin(new LoggingPlugin())
+        .build();
+```
+
 Version 2.0.0-RC5 (August 12th, 2016)
 =====================================